update src code

+++ src/featureLLDP.go
@@ -23,13 +23,13 @@ type LLDPResultType struct {
 }
 
 type PFCType struct {
-	PFCMaxClasses      uint8
-	PFCPriorityEnabled map[uint8]uint8
+	PFCMaxClasses      int
+	PFCPriorityEnabled map[int]int
 }
 
 type ETSType struct {
-	ETSTotalPG  uint8
-	ETSBWbyPGID map[uint8]uint8
+	ETSTotalPG  int
+	ETSBWbyPGID map[int]int
 }
 
 func (l *LLDPResultType) decodeLLDPPacket(packet gopacket.Packet) {
@@ -44,24 +44,24 @@ func (l *LLDPResultType) decodeLLDPPacket(packet gopacket.Packet) {
 		for _, v := range LLDPType.Values {
 			// Subtype: Priority Flow Control Configuration 0x0b
 			if v.Length == 6 && v.Value[3] == 11 {
-				l.PFC.PFCMaxClasses = uint8(bytesToDec(v.Value[4:5]))
-				PFCStatusDec := int(bytesToDec(v.Value[5:]))
+				l.PFC.PFCMaxClasses = bytesToDec(v.Value[4:5])
+				PFCStatusDec := bytesToDec(v.Value[5:])
 				l.PFC.PFCPriorityEnabled = postProPFCStatus(PFCStatusDec)
 			}
 
 			//Subtype: Port VLAN ID 0x01
 			if v.Length == 6 && v.Value[3] == 1 {
-				l.VLANID = int(bytesToDec(v.Value[4:]))
+				l.VLANID = bytesToDec(v.Value[4:])
 			}
 
 			//Subtype: ETS Recommendation 0x0a
 			if v.Length == 25 && v.Value[3] == 10 {
 				PGIDs := hex.EncodeToString(v.Value[5:9])
-				BWbyPGID := make(map[uint8]uint8)
+				BWbyPGID := make(map[int]int)
 				if len(PGIDs) != 0 {
-					l.ETS.ETSTotalPG = uint8(len(PGIDs))
+					l.ETS.ETSTotalPG = len(PGIDs)
 					for i := 0; i < 8; i++ {
-						BWbyPGID[uint8(i)] = v.Value[9+i]
+						BWbyPGID[i] = int(v.Value[9+i])
 					}
 					l.ETS.ETSBWbyPGID = BWbyPGID
 				}
@@ -118,7 +118,7 @@ func (o *OutputType) LLDPResultValidation(l *LLDPResultType, i *INIType) {
 		restultFail = append(restultFail, errMsg)
 	}
 
-	if l.ETS.ETSTotalPG != uint8(i.ETSMaxClass) {
+	if l.ETS.ETSTotalPG != i.ETSMaxClass {
 		errMsg := fmt.Sprintf("%s - Input:%d, Found: %d", WRONG_LLDP_ETS_MAX_CLASSES, i.ETSMaxClass, l.ETS.ETSTotalPG)
 		restultFail = append(restultFail, errMsg)
 	}
@@ -127,7 +127,7 @@ func (o *OutputType) LLDPResultValidation(l *LLDPResultType, i *INIType) {
 		errMsg := fmt.Sprintf("%s:\n \t\tInput:%s\n \t\tFound: %s", WRONG_LLDP_ETS_BW, i.ETSBWbyPG, etsBWString)
 		restultFail = append(restultFail, errMsg)
 	}
-	if l.PFC.PFCMaxClasses != uint8(i.PFCMaxClass) {
+	if l.PFC.PFCMaxClasses != i.PFCMaxClass {
 		errMsg := fmt.Sprintf("%s - Input:%d, Found: %d", WRONG_LLDP_PFC_MAX_CLASSES, i.PFCMaxClass, l.PFC.PFCMaxClasses)
 		restultFail = append(restultFail, errMsg)
 	}
@@ -144,28 +144,28 @@ func (o *OutputType) LLDPResultValidation(l *LLDPResultType, i *INIType) {
 	}
 }
 
-func postProPFCStatus(decNum int) map[uint8]uint8 {
+func postProPFCStatus(decNum int) map[int]int {
 	binary := 0
 	bit := 1
 	remainder := 0
-	pfcStatus := make(map[uint8]uint8, 8)
+	pfcStatus := make(map[int]int, 8)
 
 	for i := 0; i < 8; i++ {
 		remainder = decNum % 2
 		decNum = decNum / 2
 		binary += remainder * bit
 		bit *= 10
-		pfcStatus[uint8(i)] = uint8(remainder)
+		pfcStatus[i] = remainder
 	}
 
 	return pfcStatus
 }
 
-func mapintToSlicestring(mapint map[uint8]uint8) string {
+func mapintToSlicestring(mapint map[int]int) string {
 	var outputStringSlice []string
 
 	for i := 0; i < 8; i++ {
-		outputStringSlice = append(outputStringSlice, fmt.Sprintf("%d:%d", i, mapint[uint8(i)]))
+		outputStringSlice = append(outputStringSlice, fmt.Sprintf("%d:%d", i, mapint[i]))
 	}
 
 	return strings.Join(outputStringSlice, ",")
